Move some resources from common-templates to data-sources operand
These resources were moved:
- ViewRole
- EditRole
- ViewRoleBinding
- GoldenImageNamespace

Data-sources operand will contain DataImportCrons in a future patch.
