+++ internal/operands/data-sources/reconcile.go
@@ -1,0 +1,118 @@
+package data_sources
+
+import (
+	core "k8s.io/api/core/v1"
+	rbac "k8s.io/api/rbac/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	ssp "kubevirt.io/ssp-operator/api/v1beta1"
+	"kubevirt.io/ssp-operator/internal/common"
+	"kubevirt.io/ssp-operator/internal/operands"
+)
+
+// Define RBAC rules needed by this operand:
+// +kubebuilder:rbac:groups=core,resources=namespaces,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=rbac.authorization.k8s.io,resources=clusterroles;roles;rolebindings,verbs=get;list;watch;create;update;patch;delete
+
+// RBAC for created roles
+// +kubebuilder:rbac:groups=core,resources=persistentvolumeclaims,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=core,resources=persistentvolumeclaims/status,verbs=get;list;watch
+// +kubebuilder:rbac:groups=cdi.kubevirt.io,resources=datavolumes,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=cdi.kubevirt.io,resources=datavolumes/source,verbs=create
+
+const (
+	operandName      = "data-sources"
+	operandComponent = common.AppComponentTemplating
+)
+
+func WatchClusterTypes() []client.Object {
+	return []client.Object{
+		&rbac.ClusterRole{},
+		&rbac.Role{},
+		&rbac.RoleBinding{},
+		&core.Namespace{},
+	}
+}
+
+type dataSources struct{}
+
+var _ operands.Operand = &dataSources{}
+
+func New() operands.Operand {
+	return &dataSources{}
+}
+
+func (d *dataSources) Name() string {
+	return operandName
+}
+
+func (d *dataSources) WatchTypes() []client.Object {
+	return nil
+}
+
+func (d *dataSources) WatchClusterTypes() []client.Object {
+	return WatchClusterTypes()
+}
+
+func (d *dataSources) Reconcile(request *common.Request) ([]common.ReconcileResult, error) {
+	funcs := []common.ReconcileFunc{
+		reconcileGoldenImagesNS,
+		reconcileViewRole,
+		reconcileViewRoleBinding,
+		reconcileEditRole,
+	}
+	return common.CollectResourceStatus(request, funcs...)
+}
+
+func (d *dataSources) Cleanup(request *common.Request) ([]common.CleanupResult, error) {
+	return common.DeleteAll(request,
+		newGoldenImagesNS(ssp.GoldenImagesNSname),
+		newViewRole(ssp.GoldenImagesNSname),
+		newViewRoleBinding(ssp.GoldenImagesNSname),
+		newEditRole(),
+	)
+}
+
+func reconcileGoldenImagesNS(request *common.Request) (common.ReconcileResult, error) {
+	return common.CreateOrUpdate(request).
+		ClusterResource(newGoldenImagesNS(ssp.GoldenImagesNSname)).
+		WithAppLabels(operandName, operandComponent).
+		Reconcile()
+}
+
+func reconcileViewRole(request *common.Request) (common.ReconcileResult, error) {
+	return common.CreateOrUpdate(request).
+		ClusterResource(newViewRole(ssp.GoldenImagesNSname)).
+		WithAppLabels(operandName, operandComponent).
+		UpdateFunc(func(newRes, foundRes client.Object) {
+			foundRole := foundRes.(*rbac.Role)
+			newRole := newRes.(*rbac.Role)
+			foundRole.Rules = newRole.Rules
+		}).
+		Reconcile()
+}
+
+func reconcileViewRoleBinding(request *common.Request) (common.ReconcileResult, error) {
+	return common.CreateOrUpdate(request).
+		ClusterResource(newViewRoleBinding(ssp.GoldenImagesNSname)).
+		WithAppLabels(operandName, operandComponent).
+		UpdateFunc(func(newRes, foundRes client.Object) {
+			newBinding := newRes.(*rbac.RoleBinding)
+			foundBinding := foundRes.(*rbac.RoleBinding)
+			foundBinding.Subjects = newBinding.Subjects
+			foundBinding.RoleRef = newBinding.RoleRef
+		}).
+		Reconcile()
+}
+
+func reconcileEditRole(request *common.Request) (common.ReconcileResult, error) {
+	return common.CreateOrUpdate(request).
+		ClusterResource(newEditRole()).
+		WithAppLabels(operandName, operandComponent).
+		UpdateFunc(func(newRes, foundRes client.Object) {
+			newRole := newRes.(*rbac.ClusterRole)
+			foundRole := foundRes.(*rbac.ClusterRole)
+			foundRole.Rules = newRole.Rules
+		}).
+		Reconcile()
+}
